added and limited access controls for user and guest

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,19 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]); // This constructor blocks
+    }                                                           // unregistered users being able
+                                                                // to create new posts. It allows 
+                                                                // the index and show pages to be 
+                                                                // seen with the 'except' method.
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -76,6 +89,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for user, if not post creator, redirect to posts and display error message.
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Access');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -111,6 +130,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        // Check for user, if not post creator, redirect to posts and display error message.
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Access');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
